Move partition logic to client package.
This simplifies automatic updates to the client.

Existing imports of the form `from aryn_sdk.partition import partition_file`
will continue to work.
